requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.35
pydantic>=2.9.2
pydantic-settings>=2.5.2
httpx>=0.27.2
tenacity>=9.0.0
numpy>=2.1.0
faiss-cpu>=1.13.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-mistralai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
streamlit>=1.39.0
python-multipart>=0.0.12
pypdf>=5.4.0
python-docx>=1.1.2
pytest>=8.3.3
pytest-asyncio>=0.24.0
